@@ -13,11 +13,22 @@ import (
 
 func (s AsertoPluginServer) Import(srv proto.Plugin_ImportServer) error {
 	errc := make(chan error, 128)
-	errDone := make(chan bool, 1)
+	errDone := make(chan struct{}, 1)
+	pluginClosed := false
 
 	defer func() {
+		close(errc)
+		<-errDone
+
+		if !pluginClosed {
+			_, err := s.PluginHandler.Close()
+			if err != nil {
+				log.Println(err.Error())
+			}
+		}
+
 		if r := recover(); r != nil {
-			errc <- fmt.Errorf("recovering from panic in Import error is: %v", r)
+			log.Println(fmt.Errorf("recovering from panic in Import error is: %v", r))
 		}
 	}()
 
@@ -25,11 +36,11 @@ func (s AsertoPluginServer) Import(srv proto.Plugin_ImportServer) error {
 	cfg := s.PluginHandler.GetConfig()
 
 	go func() {
+		defer close(errDone)
 		for {
 			e, more := <-errc
 			if !more {
 				// channel closed
-				errDone <- true
 				return
 			}
 			err := srv.Send(
@@ -41,6 +52,7 @@ func (s AsertoPluginServer) Import(srv proto.Plugin_ImportServer) error {
 			)
 			if err != nil {
 				log.Println(err.Error())
+				return
 			}
 		}
 	}()
@@ -52,12 +64,13 @@ func (s AsertoPluginServer) Import(srv proto.Plugin_ImportServer) error {
 		}
 		if err != nil {
 			errc <- err
+			break
 		}
 
 		if !initialized {
 			err = config.NewConfig(req.GetConfig(), cfg)
 			if err != nil {
-				errc <- err
+				return err
 			}
 			err := s.PluginHandler.Open(cfg, OperationTypeWrite)
 			if err != nil {
@@ -79,6 +92,7 @@ func (s AsertoPluginServer) Import(srv proto.Plugin_ImportServer) error {
 		errc <- err
 	}
 
+	pluginClosed = true
 	if stats != nil {
 		err = srv.Send(
 			&proto.ImportResponse{
@@ -91,11 +105,9 @@ func (s AsertoPluginServer) Import(srv proto.Plugin_ImportServer) error {
 				}},
 		)
 		if err != nil {
-			errc <- err
+			return err
 		}
 	}
 
-	close(errc)
-	<-errDone
 	return nil
 }
